+++ test/flop_phoenix_test.exs
@@ -120,6 +120,7 @@ defmodule Flop.PhoenixTest do
 
       a = find_one(html, "a:fl-contains('Previous')")
 
+      assert attribute(a, "rel") == "prev"
       assert attribute(a, "class") == "pagination-previous"
       assert attribute(a, "data-phx-link") == "patch"
       assert attribute(a, "data-phx-link-state") == "push"
@@ -139,6 +140,7 @@ defmodule Flop.PhoenixTest do
 
       button = find_one(html, "button:fl-contains('Previous')")
 
+      assert attribute(button, "rel") == nil
       assert attribute(button, "class") == "pagination-previous"
       assert attribute(button, "data-phx-link") == nil
       assert attribute(button, "data-phx-link-state") == nil
@@ -162,6 +164,7 @@ defmodule Flop.PhoenixTest do
 
       a = find_one(html, "a:fl-contains('Previous')")
 
+      assert attribute(a, "rel") == "prev"
       assert attribute(a, "class") == "pagination-previous"
       assert attribute(a, "data-phx-link") == "patch"
       assert attribute(a, "data-phx-link-state") == "push"
@@ -227,6 +230,7 @@ defmodule Flop.PhoenixTest do
 
       assert attribute(button, "phx-value-page") == "1"
       assert attribute(button, "href") == nil
+      assert attribute(button, "rel") == nil
     end
 
     test "adds phx-target to previous button" do
@@ -325,6 +329,7 @@ defmodule Flop.PhoenixTest do
       assert attribute(previous_link, "aria-disabled") == "true"
       assert attribute(previous_link, "href") == nil
       assert attribute(previous_link, "role") == "link"
+      assert attribute(previous_link, "rel") == nil
     end
 
     test "disables previous link if on first page when using on_paginate" do
@@ -335,11 +340,11 @@ defmodule Flop.PhoenixTest do
         <Flop.Phoenix.pagination meta={@meta} on_paginate={JS.push("paginate")} />
         """)
 
-      assert previous_link = find_one(html, "a:fl-contains('Previous')")
+      assert button = find_one(html, "button:fl-contains('Previous')")
 
-      assert attribute(previous_link, "class") == "pagination-previous"
-      assert attribute(previous_link, "href") == nil
-      assert attribute(previous_link, "aria-disabled") == "true"
+      assert attribute(button, "class") == "pagination-previous"
+      assert attribute(button, "href") == nil
+      assert attribute(button, "disabled") == "disabled"
     end
 
     test "allows to overwrite previous link class and content if disabled" do
@@ -459,7 +464,7 @@ defmodule Flop.PhoenixTest do
       assert attribute(next, "role") == "link"
     end
 
-    test "renders next link on last page when using on_paginate" do
+    test "renders disabled next button on last page when using on_paginate" do
       assigns = %{meta: build(:meta_on_last_page)}
 
       html =
@@ -467,11 +472,10 @@ defmodule Flop.PhoenixTest do
         <Flop.Phoenix.pagination meta={@meta} on_paginate={JS.push("paginate")} />
         """)
 
-      assert next = find_one(html, "a:fl-contains('Next')")
+      assert next = find_one(html, "button:fl-contains('Next')")
       assert attribute(next, "class") == "pagination-next"
       assert attribute(next, "href") == nil
-      assert attribute(next, "role") == "link"
-      assert attribute(next, "aria-disabled") == "true"
+      assert attribute(next, "disabled") == "disabled"
     end
 
     test "allows to overwrite next link attributes and content when disabled" do
@@ -703,6 +707,7 @@ defmodule Flop.PhoenixTest do
       assert attribute(previous, "class") == "pagination-previous"
       assert attribute(previous, "data-phx-link") == "patch"
       assert attribute(previous, "data-phx-link-state") == "push"
+      assert attribute(previous, "rel") == "prev"
       assert href = attribute(previous, "href")
       assert_urls_match(href, "/pets", expected_query.(1))
 
@@ -717,6 +722,7 @@ defmodule Flop.PhoenixTest do
       assert attribute(next, "class") == "pagination-next"
       assert attribute(next, "data-phx-link") == "patch"
       assert attribute(next, "data-phx-link-state") == "push"
+      assert attribute(next, "rel") == "next"
       assert href = attribute(next, "href")
       assert_urls_match(href, "/pets", expected_query.(3))
     end
@@ -1112,6 +1118,7 @@ defmodule Flop.PhoenixTest do
       assert attribute(button, "data-phx-link") == nil
       assert attribute(button, "data-phx-link-state") == nil
       assert attribute(button, "href") == nil
+      assert attribute(button, "rel") == nil
       assert attribute(button, "phx-value-to") == "previous"
       assert phx_click = attribute(button, "phx-click")
       assert Jason.decode!(phx_click) == [["push", %{"event" => "paginate"}]]
@@ -1134,6 +1141,7 @@ defmodule Flop.PhoenixTest do
       assert attribute(link, "class") == "pagination-previous"
       assert attribute(link, "data-phx-link") == "patch"
       assert attribute(link, "data-phx-link-state") == "push"
+      assert attribute(link, "rel") == "prev"
       assert href = attribute(link, "href")
       assert_urls_match(href, "/pets?before=B&last=10")
 
@@ -1200,6 +1208,7 @@ defmodule Flop.PhoenixTest do
 
       assert attribute(button, "phx-value-to") == "previous"
       assert attribute(button, "href") == nil
+      assert attribute(button, "rel") == nil
     end
 
     test "adds phx-target to previous button" do
@@ -1228,9 +1237,11 @@ defmodule Flop.PhoenixTest do
         """)
 
       assert link = find_one(html, "a:fl-contains('Previous')")
+      assert attribute(link, "rel") == "prev"
       assert href = attribute(link, "href")
       assert_urls_match(href, "/pets?last=10&before=B")
       assert link = find_one(html, "a:fl-contains('Next')")
+      assert attribute(link, "rel") == "next"
       assert href = attribute(link, "href")
       assert_urls_match(href, "/pets?first=10&after=C")
 
@@ -1241,10 +1252,12 @@ defmodule Flop.PhoenixTest do
         """)
 
       assert link = find_one(html, "a:fl-contains('Previous')")
+      assert attribute(link, "rel") == "prev"
       assert href = attribute(link, "href")
       assert_urls_match(href, "/pets?first=10&after=C")
 
       assert link = find_one(html, "a:fl-contains('Next')")
+      assert attribute(link, "rel") == "next"
       assert href = attribute(link, "href")
       assert_urls_match(href, "/pets?last=10&before=B")
     end
@@ -1307,9 +1320,10 @@ defmodule Flop.PhoenixTest do
 
       assert attribute(previous_link, "class") == "pagination-previous"
       assert attribute(previous_link, "aria-disabled") == "true"
+      assert attribute(previous_link, "rel") == nil
     end
 
-    test "disables previous link if on first page when using on_paginate" do
+    test "disables previous button if on first page when using on_paginate" do
       assigns = %{meta: build(:meta_with_cursors, has_previous_page?: false)}
 
       html =
@@ -1317,12 +1331,12 @@ defmodule Flop.PhoenixTest do
         <Flop.Phoenix.pagination meta={@meta} on_paginate={JS.push("paginate")} />
         """)
 
-      assert previous_link = find_one(html, "a:fl-contains('Previous')")
+      assert previous_button = find_one(html, "button:fl-contains('Previous')")
 
-      assert attribute(previous_link, "class") == "pagination-previous"
-      assert attribute(previous_link, "href") == nil
-      assert attribute(previous_link, "aria-disabled") == "true"
-      assert attribute(previous_link, "role") == "link"
+      assert attribute(previous_button, "class") == "pagination-previous"
+      assert attribute(previous_button, "href") == nil
+      assert attribute(previous_button, "disabled") == "disabled"
+      assert attribute(previous_button, "rel") == nil
     end
 
     test "allows to overwrite previous link class and content if disabled" do
@@ -1340,13 +1354,12 @@ defmodule Flop.PhoenixTest do
         />
         """)
 
-      assert previous_link = find_one(html, "a:fl-contains('Prev')")
-      assert attribute(previous_link, "class") == "prev"
-      assert attribute(previous_link, "title") == "no"
-      assert attribute(previous_link, "href") == nil
-      assert attribute(previous_link, "aria-disabled") == "true"
-      assert attribute(previous_link, "role") == "link"
-      assert text(previous_link) == "Prev"
+      assert previous_button = find_one(html, "button:fl-contains('Prev')")
+      assert attribute(previous_button, "class") == "prev"
+      assert attribute(previous_button, "title") == "no"
+      assert attribute(previous_button, "href") == nil
+      assert attribute(previous_button, "disabled") == "disabled"
+      assert text(previous_button) == "Prev"
     end
 
     test "renders next link" do
@@ -1439,7 +1452,7 @@ defmodule Flop.PhoenixTest do
       assert attribute(next, "aria-disabled") == "true"
     end
 
-    test "renders next link on last page when using on_paginate" do
+    test "renders disabled next button on last page when using on_paginate" do
       assigns = %{meta: build(:meta_with_cursors, has_next_page?: false)}
 
       html =
@@ -1447,14 +1460,13 @@ defmodule Flop.PhoenixTest do
         <Flop.Phoenix.pagination meta={@meta} on_paginate={JS.push("paginate")} />
         """)
 
-      assert next = find_one(html, "a:fl-contains('Next')")
+      assert next = find_one(html, "button:fl-contains('Next')")
       assert attribute(next, "class") == "pagination-next"
-      assert attribute(next, "aria-disabled") == "true"
-      assert attribute(next, "role") == "link"
+      assert attribute(next, "disabled") == "disabled"
       assert attribute(next, "href") == nil
     end
 
-    test "allows to overwrite next link attributes and content when disabled" do
+    test "allows to overwrite next button attributes and content when disabled" do
       assigns = %{meta: build(:meta_with_cursors, has_next_page?: false)}
 
       html =
@@ -1469,11 +1481,11 @@ defmodule Flop.PhoenixTest do
         />
         """)
 
-      assert next_link = find_one(html, "a:fl-contains('N-n-next')")
-      assert attribute(next_link, "class") == "next"
-      assert attribute(next_link, "aria-disabled") == "true"
-      assert attribute(next_link, "href") == nil
-      assert attribute(next_link, "title") == "no"
+      assert next_button = find_one(html, "button:fl-contains('N-n-next')")
+      assert attribute(next_button, "class") == "next"
+      assert attribute(next_button, "disabled") == "disabled"
+      assert attribute(next_button, "href") == nil
+      assert attribute(next_button, "title") == "no"
     end
 
     test "raises if neither path nor event are passed" do